refactor: Tooltip 컴포넌트 수정 및 스토리 추가
