+++ src/components/common/Tooltip/style.css
@@ -1,0 +1,22 @@
+[data-scope="tooltip"][data-part="trigger"] {
+  position: relative;
+  cursor: help;
+}
+
+[data-scope="tooltip"][data-part="content"] {
+  position: absolute;
+  top: 50%;
+  left: 0;
+  transform: translateX(25%);
+  width: 125px;
+  margin-top: 5px;
+  font-size: 12px;
+  word-break: keep-all;
+  background-color: white;
+  box-shadow:
+    rgba(50, 50, 93, 0.25) 0px 2px 5px -1px,
+    rgba(0, 0, 0, 0.3) 0px 1px 3px -1px;
+  padding: 5px;
+  border-radius: 3px;
+  z-index: 1000;
+}
